IO folder with refactored config reader.
